pyproject.toml
[project]
name = "corporate-reports"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "stf", email = "123456789+stf@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
